move subcat slider functions to their own file
